Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   edge.py
	modified:   utils/contra.py

+++ utils/contra.py
@@ -30,7 +30,7 @@ def contra(args):
             # a_inv = modinv(args.a[i], args.p)
             # g_inv = modinv(args.g, args.p)
             # last_layer = ((args.cos_client_ref[i] * a_inv) % args.p * g_inv) % args.p
-            last_layer = (args.cos_client_ref[i].long() * a_inv) % args.p
+            last_layer = (args.cos_client_ref[i] * a_inv) % args.p
             last_layer[last_layer > args.p // 2] -= args.p      
             last_layer = uncast_from_range(last_layer, args.g)
             if i == 5:
